examples/python/projector-receipt/pyproject.toml
[project]
name = "projector-receipt"
version = "0.1.0"
description = "Receipt Projector gRPC server for angzarr examples"
requires-python = ">=3.11"
dependencies = [
    "protobuf>=5.0.0",
    "grpcio>=1.60.0",
    "grpcio-health-checking>=1.60.0",
    "structlog>=24.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-bdd>=7.0.0",
    "debugpy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
